refactor: Add Tauri build configuration and update dependencies
This commit adds the necessary Tauri build configuration files and updates the dependencies in the project. The `src-tauri/build.rs` file is added to handle the Tauri build process. The `.gitignore` file in the `src-tauri` directory is updated to exclude the `/target/` directory, which contains compiled files and executables. The `package.json` file is updated to include the `@tauri-apps/cli` dependency version `1.6.0`. The `next.config.mjs` file is updated to set the `distDir` option to "build". The `src-tauri/src/main.rs` file is added to handle the Tauri application initialization. The `src-tauri/Cargo.toml` file is added to specify the Tauri app's name, version, and dependencies. The `src/pages/album/[id].tsx` file is updated to use the first artist in the `artists` array. The `src/contexts/MusicPlayerContext.jsx` file is updated to use the first artist in the `artists` array for the media session metadata. The `src/util/signin.ts` file is updated to use the `response` property of the API response data. The `src/util/verify.ts` file is updated to handle errors and use the `response` property of the API response data. The `src/pages/auth/verify.tsx` file is updated to handle null values in the `responseMessage` state. The `src/pages/auth/signup.tsx` and `src/pages/auth/signin.tsx` files are updated to remove unnecessary console.log statements. The `src/pages/admin/add/track.tsx` file is updated to check the `status` property of the API response data. The `src-tauri/src/main.rs` and `src-tauri/Cargo.toml` files are added to the repository.